requirements.txt
streamlit>=1.37.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.1
openai>=1.40.0
google-genai>=0.3.0